renamed dsclient
