+++ src/PluginsTreeView.js
@@ -1,5 +1,5 @@
 const vscode = require('vscode');
-const ext = require("./extension");
+const ext = require("./dsclient");
 
 class TreeDataProvider {
