+++ src/SamplesTreeView.js
@@ -1,5 +1,5 @@
 const vscode = require('vscode');
-const ext = require("./extension");
+const ext = require("./dsclient");
 const localData = require('./local-data');
 
 class TreeDataProvider {
