+++ src/commands/delete-app.js
@@ -1,5 +1,5 @@
 const vscode = require('vscode');
-const ext = require('../extension');
+const ext = require('../dsclient');
 const rimraf = require("rimraf");
